backend/src/requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.2
Automat==20.2.0
awsebcli==3.19.4
bcrypt==3.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
blessed==1.18.0
boto==2.49.0
boto3==1.17.27
botocore==1.20.39
bs4==0.0.1
cached-property==1.5.2
celery==5.0.5
cement==2.8.2
certifi==2020.12.5
cffi==1.14.4
channels==3.0.2
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
defusedxml==0.6.0
dj-rest-auth==2.1.2
Django==3.1.4
django-allauth==0.44.0
django-cors-headers==3.6.0
django-private-storage==2.2.2
django-redis==4.12.1
django-referrer-policy==1.0
django-rest-auth==0.9.5
django-storages==1.11.1
django-tgbot==1.4.9
djangorestframework==3.12.2
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
h2==3.2.0
hiredis==1.1.0
hpack==3.0.0
http-ece==1.1.0
hyperframe==5.2.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==3.7.3
incremental==17.5.0
jmespath==0.10.0
jsonschema==3.2.0
kombu==5.0.2
msgpack==1.0.2
oauthlib==3.1.0
paramiko==2.7.2
pathspec==0.5.9
Pillow==8.1.0
priority==1.3.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
py-vapid==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycurl==7.43.0.5
PyHamcrest==2.0.2
PyJWT==1.7.1
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
pywebpush
PyYAML==5.3.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.6
semantic-version==2.8.5
service-identity==18.1.0
setuptools-rust==0.12.1
six
soupsieve==2.1
sqlparse==0.4.1
termcolor==1.1.0
texttable==1.6.3
toml==0.10.2
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
vine==5.0.0
wcwidth==0.1.9
websocket-client==0.58.0
zipp==3.4.1
zope.interface==5.2.0
